+++ src/main/java/services/UserService.java
@@ -13,7 +13,7 @@ public class UserService {
 		return DAOFactory.getUserDAO().findAll();
 	}
 
-	public User create(String username, String password, Integer coins, Double time) {
+	public User create(String username, String password, Double coins, Double time) {
 		User user = new User(-1, username, password, coins, time, false);
 		user.setPassword(password);
